refactor(modelprovider): Update component props in created hook for Nuxt SSR
For Nuxt SSR the component props need to be set in the created instead of the mounted hook (which is not executed on the server). For the props to be actually available in render() though, the model has to be fetched directly (and synchronously) from the ModelStore. Otherwise the promise returned by ModelManager.getData() will only be resolved after rendering.
For this to work the model data needs to be provided to the ModelManager before the components are rendered. A better solution would be to put the model retrieval into the Nuxt fetch() hook (which blocks rendering), but this caused serialization errors.